// packages/form/lib/types.ts
export type ParsePathKey<T> = ParsePath<T, keyof T> | keyof T;

export type Path<T> = ParsePathKey<T> extends string | keyof T ? ParsePathKey<T> : keyof T;

export type PathValue<T, P extends Path<T>> =
  P extends `${infer Key}.${infer Rest}`
  ? Key extends keyof T ? Rest extends Path<T[Key]>
  ? PathValue<T[Key], Rest>
  : never
  : never
  : P extends keyof T ? T[P]
  : never;


export type AnyString = string & { value?: any };
export type KeyOfAny<K extends string | number | symbol> = AnyString | K;

export type DeepPartial<T> = {
  [P in keyof T]: DeepPartial<T[P]>;
};

export type FlatMap<T extends Record<string, unknown>> =
  { [P in Path<T>]: PathValue<T, P> }
// ...
export type FormElement =
  | HTMLInputElement
  | HTMLTextAreaElement
  | HTMLSelectElement;

export type ElementType = keyof typeof TYPE_MAP;

export type DataTypeBase = 'string' | 'number' | 'integer' | 'float' | 'bigint' | 'boolean' | 'none';

export type DataType = DataTypeBase | `[${Exclude<DataTypeBase, 'none'>}]`;